# Backend2_assignment-1/api/src/main/java/com/example/api/dto/CustomerItemResponseDTO.java
package com.example.api.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CustomerItemResponseDTO {
    private Long customerId;
    private Long itemId;
}