# Backend2_assignment-1/api/src/main/java/com/example/api/model/Customer.java
package com.example.api.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;

@Entity
@Data       // Lombok (innehåller @NoArgsConstructor)
@NoArgsConstructor
@AllArgsConstructor
@Builder    // Build/create customer
public class Customer {

    @Id
    @SequenceGenerator(
            name = "customer_sequence",
            sequenceName = "customer_sequence",
            allocationSize = 1
    )
    @GeneratedValue(
            strategy = GenerationType.SEQUENCE,
            generator = "customer_sequence"
    )
    private Long id;

    private String name;
    private String address;
    private String username;
    private String password;
}